Decouple core transaction object from zerv transaction implementation to promote reuse outside  zerv environment
